refactor add card component
