+++ src/pages/Board/components/InnerList/types.ts
